move bounding box to geometry module
