+++ include/core/rasterizer.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "Ops.h"
-#include "core/bounding_box.hpp"
+#include "geometry//bounding_box.hpp"
 #include "core/camera.hpp"
 #include "core/splat_data.hpp"
 #include <torch/torch.h>
